Move code into scigateway_auth package
