+++ scigateway_auth/common/exceptions.py
